Move QuickPick on to an ObjectTree (#207520)
* Move QuickPick on to an ObjectTree

Enables sticky scroll

* Fix tests

* fix tests take 2

* fix tests by cleaning up disposables

* fix tests again

* `setChildren` less

* test if this is what fixes CI (but local would be broken)

* remove setFocus call since it just sends a random event that we don't need

* Only apply events to quickpickitems not separators

* a comment
