+++ src/vs/platform/quickinput/test/browser/quickinput.test.ts
@@ -6,9 +6,9 @@
 import * as assert from 'assert';
 import { unthemedInboxStyles } from 'vs/base/browser/ui/inputbox/inputBox';
 import { unthemedButtonStyles } from 'vs/base/browser/ui/button/button';
-import { IListRenderer, IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
-import { IListOptions, List, unthemedListStyles } from 'vs/base/browser/ui/list/listWidget';
+import { unthemedListStyles } from 'vs/base/browser/ui/list/listWidget';
 import { unthemedToggleStyles } from 'vs/base/browser/ui/toggle/toggle';
+import { Event } from 'vs/base/common/event';
 import { raceTimeout } from 'vs/base/common/async';
 import { unthemedCountStyles } from 'vs/base/browser/ui/countBadge/countBadge';
 import { unthemedKeybindingLabelOptions } from 'vs/base/browser/ui/keybindingLabel/keybindingLabel';
@@ -20,6 +20,18 @@ import { toDisposable } from 'vs/base/common/lifecycle';
 import { mainWindow } from 'vs/base/browser/window';
 import { QuickPick } from 'vs/platform/quickinput/browser/quickInput';
 import { IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
+import { ILayoutService } from 'vs/platform/layout/browser/layoutService';
+import { IContextViewService } from 'vs/platform/contextview/browser/contextView';
+import { IListService, ListService } from 'vs/platform/list/browser/listService';
+import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
+import { ContextKeyService } from 'vs/platform/contextkey/browser/contextKeyService';
+import { NoMatchingKb } from 'vs/platform/keybinding/common/keybindingResolver';
+import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
+import { ContextViewService } from 'vs/platform/contextview/browser/contextViewService';
 
 // Sets up an `onShow` listener to allow us to wait until the quick pick is shown (useful when triggering an `accept()` right after launching a quick pick)
 // kick this off before you launch the picker and then await the promise returned after you launch the picker.
@@ -45,50 +57,58 @@ suite('QuickInput', () => { // https://github.com/microsoft/vscode/issues/147543
 		mainWindow.document.body.appendChild(fixture);
 		store.add(toDisposable(() => mainWindow.document.body.removeChild(fixture)));
 
-		controller = store.add(new QuickInputController({
-			container: fixture,
-			idPrefix: 'testQuickInput',
-			ignoreFocusOut() { return true; },
-			returnFocus() { },
-			backKeybindingLabel() { return undefined; },
-			setContextKey() { return undefined; },
-			linkOpenerDelegate(content) { },
-			createList: <T>(
-				user: string,
-				container: HTMLElement,
-				delegate: IListVirtualDelegate<T>,
-				renderers: IListRenderer<T, any>[],
-				options: IListOptions<T>,
-			) => new List<T>(user, container, delegate, renderers, options),
-			hoverDelegate: {
-				showHover(options, focus) {
-					return undefined;
+		const instantiationService = new TestInstantiationService();
+
+		// Stub the services the quick input controller needs to function
+		instantiationService.stub(IThemeService, new TestThemeService());
+		instantiationService.stub(IConfigurationService, new TestConfigurationService());
+		instantiationService.stub(IListService, store.add(new ListService()));
+		instantiationService.stub(ILayoutService, { activeContainer: fixture, onDidLayoutContainer: Event.None } as any);
+		instantiationService.stub(IContextViewService, store.add(instantiationService.createInstance(ContextViewService)));
+		instantiationService.stub(IContextKeyService, store.add(instantiationService.createInstance(ContextKeyService)));
+		instantiationService.stub(IKeybindingService, {
+			mightProducePrintableCharacter() { return false; },
+			softDispatch() { return NoMatchingKb; },
+		});
+
+		controller = store.add(instantiationService.createInstance(
+			QuickInputController,
+			{
+				container: fixture,
+				idPrefix: 'testQuickInput',
+				ignoreFocusOut() { return true; },
+				returnFocus() { },
+				backKeybindingLabel() { return undefined; },
+				setContextKey() { return undefined; },
+				linkOpenerDelegate(content) { },
+				hoverDelegate: {
+					showHover(options, focus) {
+						return undefined;
+					},
+					delay: 200
 				},
-				delay: 200
-			},
-			styles: {
-				button: unthemedButtonStyles,
-				countBadge: unthemedCountStyles,
-				inputBox: unthemedInboxStyles,
-				toggle: unthemedToggleStyles,
-				keybindingLabel: unthemedKeybindingLabelOptions,
-				list: unthemedListStyles,
-				progressBar: unthemedProgressBarOptions,
-				widget: {
-					quickInputBackground: undefined,
-					quickInputForeground: undefined,
-					quickInputTitleBackground: undefined,
-					widgetBorder: undefined,
-					widgetShadow: undefined,
-				},
-				pickerGroup: {
-					pickerGroupBorder: undefined,
-					pickerGroupForeground: undefined,
+				styles: {
+					button: unthemedButtonStyles,
+					countBadge: unthemedCountStyles,
+					inputBox: unthemedInboxStyles,
+					toggle: unthemedToggleStyles,
+					keybindingLabel: unthemedKeybindingLabelOptions,
+					list: unthemedListStyles,
+					progressBar: unthemedProgressBarOptions,
+					widget: {
+						quickInputBackground: undefined,
+						quickInputForeground: undefined,
+						quickInputTitleBackground: undefined,
+						widgetBorder: undefined,
+						widgetShadow: undefined,
+					},
+					pickerGroup: {
+						pickerGroupBorder: undefined,
+						pickerGroupForeground: undefined,
+					}
 				}
 			}
-		},
-			new TestThemeService(),
-			{ activeContainer: fixture } as any));
+		));
 
 		// initial layout
 		controller.layout({ height: 20, width: 40 }, 0);
